docker_gcp/requirements.txt
requests
tensorflow
efficientnet
streamlit
google_api_python_client
protobuf
